requirements.txt
streamlit
openai
audio_recorder_streamlit
streamlit-float
python-dotenv==1.0.0
